Refactor to accept more transactions
