+++ index.ts
@@ -1,4 +1,3 @@
-import { BigNumber } from "bignumber.js";
 import { H160, H256, U64 } from "codechain-primitives";
 import { SDK } from "codechain-sdk";
 import { Asset } from "codechain-sdk/lib/core/Asset";
@@ -9,59 +8,9 @@ import { KeyStore } from "codechain-sdk/lib/key/KeyStore";
 import { blake160 } from "codechain-sdk/lib/utils";
 import * as request from "request-promise-native";
 import * as sleep from "sleep";
-import { calculateSeq, getConfig, haveConfig, sendTransaction } from "./util";
+import { chooseAccount, getConfig, haveConfig, PayerInfo } from "./util";
 
 const airdropOilWaitingLimit = 10;
-
-interface Account {
-    address: string;
-    balance: BigNumber;
-}
-
-function getRandomAccount(accounts: Account[]): string {
-    const totalBalance = accounts.reduce(
-        (acc, account) => account.balance.plus(acc),
-        new BigNumber(0)
-    );
-    const random = new BigNumber(Math.random()).multipliedBy(totalBalance);
-    let sum = new BigNumber(0);
-
-    for (const account of accounts) {
-        sum = sum.plus(account.balance);
-        if (random.isLessThan(sum)) {
-            return account.address;
-        }
-    }
-    throw new Error("unreachable");
-}
-
-async function fetchAccounts(): Promise<Account[]> {
-    const indexerUrl = getConfig<string>("indexer_url");
-    const accountsUrl = `${indexerUrl}/api/account`;
-    const items: { address: string; balance: string }[] = await request({
-        url: accountsUrl,
-        json: true
-    });
-
-    return items.map(item => {
-        const address = item.address;
-        const balance = new BigNumber(item.balance, 10);
-        return { address, balance };
-    });
-}
-
-async function chooseAccount(
-    payer: string,
-    excludedAccountList: string[]
-): Promise<string> {
-    const accounts = (await fetchAccounts()).filter(
-        account =>
-            account.address !== payer &&
-            !account.balance.isZero() &&
-            excludedAccountList.indexOf(account.address) === -1
-    );
-    return getRandomAccount(accounts);
-}
 
 async function airdropCCCTransaction(
     sdk: SDK,
@@ -236,12 +185,11 @@ async function main() {
     const networkId = getConfig<string>("network_id");
 
     const sdk = new SDK({ server: rpcUrl, networkId });
-
     const keyStore = await sdk.key.createLocalKeyStore();
-
     const payer = getConfig<string>("payer.payer");
-
     const payerPassphrase = getConfig<string>("payer.passphrase");
+
+    const payerInfo = new PayerInfo(sdk, payer, payerPassphrase, keyStore);
     const reward = getConfig<number>("reward");
 
     const dropInterval = getConfig<number>("drop_interval");
@@ -260,15 +208,7 @@ async function main() {
                 excludedAccountList,
                 reward
             );
-            const seq = await calculateSeq(sdk, payer);
-            await sendTransaction(
-                sdk,
-                payer,
-                payerPassphrase,
-                keyStore,
-                seq,
-                transaction
-            );
+            await payerInfo.sendTransaction(transaction);
             console.log("CCC is airdropped");
         } catch (err) {
             console.error(err);
@@ -306,13 +246,7 @@ async function main() {
                     keyStore,
                     dropInterval
                 );
-                const seq = await calculateSeq(sdk, payer);
-                const sentOilTransactionHash = await sendTransaction(
-                    sdk,
-                    payer,
-                    payerPassphrase,
-                    keyStore,
-                    seq,
+                const sentOilTransactionHash = await payerInfo.sendTransaction(
                     oilTransaction
                 );
                 console.log(
